sci-paper-automation/requirements.txt
requests>=2.32.0
pydantic>=2.10.0
PyYAML>=6.0.2
python-docx>=1.1.2
pandas>=2.2.3
matplotlib>=3.10.0
seaborn>=0.13.2
biopython>=1.85
arxiv>=2.1.3
